Enhancement: Adds fallback to default memomry configuration of 1g min and 4g max for docker containers with no set environment variables for memory size of Structr service.

+++ structr-ui/src/main/resources/bin/docker.sh
@@ -42,17 +42,22 @@ function addEnvironmentEntryToConfig {
 	echo "${SETTING}=${VALUE}" >> "${_STRUCTR_HOME}"/structr.conf
 }
 
-# create memory config file before startup
-touch "${_STRUCTR_HOME}/bin/memory.config"
-if [ -n "$STRUCTR_MAX_HEAP" ] && ! [ -n "$STRUCTR_MIN_HEAP" ]; then
-    echo "STRUCTR_MAX_HEAP set but STRUCTR_MIN_HEAP not found in docker environment"
-    exit 1;
-  elif [ -n "$STRUCTR_MIN_HEAP" ] && ! [ -n "$STRUCTR_MAX_HEAP" ]; then
-    echo "STRUCTR_MIN_HEAP set but STRUCTR_MIN_HEAP not found in docker environment"
-    exit 1;
-  else
-    echo "Creating memory configuration file with -Xms${STRUCTR_MIN_HEAP} -Xmx${STRUCTR_MAX_HEAP}"
-    echo "-Xms${STRUCTR_MIN_HEAP} -Xmx${STRUCTR_MAX_HEAP}" > "${_STRUCTR_HOME}/bin/memory.config"
+# check and create memory config file before startup
+if ! [ -f "${_STRUCTR_HOME}/bin/memory.config" ]; then
+  touch "${_STRUCTR_HOME}/bin/memory.config"
+  if [ -n "$STRUCTR_MAX_HEAP" ] && ! [ -n "$STRUCTR_MIN_HEAP" ]; then
+      echo "STRUCTR_MAX_HEAP set but STRUCTR_MIN_HEAP not found in docker environment"
+      exit 1;
+    elif [ -n "$STRUCTR_MIN_HEAP" ] && ! [ -n "$STRUCTR_MAX_HEAP" ]; then
+      echo "STRUCTR_MIN_HEAP set but STRUCTR_MIN_HEAP not found in docker environment"
+      exit 1;
+    elif ! [ -n "$STRUCTR_MIN_HEAP" ] && ! [ -n "$STRUCTR_MAX_HEAP" ]; then
+        echo "Creating memory configuration file with default configuration of -Xms1g -Xmx4g"
+        echo "-Xms1g -Xmx4g" > "${_STRUCTR_HOME}/bin/memory.config"
+    else
+      echo "Creating memory configuration file with -Xms${STRUCTR_MIN_HEAP} -Xmx${STRUCTR_MAX_HEAP}"
+      echo "-Xms${STRUCTR_MIN_HEAP} -Xmx${STRUCTR_MAX_HEAP}" > "${_STRUCTR_HOME}/bin/memory.config"
+  fi
 fi
 
 unset STRUCTR_MIN_HEAP
